Referential transparency
Testing is much easier if `updateQuality()` is referentially transparent.
Rather than rewrite all of the original code,
I've introduced a small change that copies the item before updating it.
This is a cheap way to make everything testable
without risking changing the semantics.

Also, as `updateQuality()` applies the same transform to every item in the list,
I've introduced a second version of the method that updates a single `Item`.
This will be simpler to test than the multi-`Item` version.